backend god (not really)
added in a history in the console, just pulls and writes to a text file. Also added it in as a command in the terminal. Also started dev on a push command, this allows you to push a command to all hosts, it is close to working but it has something wrong atm, just doesnt seem to call the function when push is called not sure why yet