server-communication
added in stuff to imporve the communication between the server and client, can better execute commands now. Issue seems to be the keepalive is being sent as the response to the server compared to what should be pulled. Also added in a systemd service on the network.rs file, this is so once the binary is run as root it will auto create a systemd service to keep it alive