not really working
tried to move these functions to taraia so they could be module to use in the frontend aswell. ran into issues with the shared state. once i was able to get it compilied there were still issues, the agent could connect to the server but the console didnt show it as a registered agent